+++ examples/RP2040_WiFiNINA_MQTT/defines.h
@@ -162,6 +162,11 @@
 
 /////////////////////////////////////////////
 
+// Optional, to use Board Name in Menu
+#define USING_BOARD_NAME                    true
+
+/////////////////////////////////////////////
+
 #include <WiFiManager_NINA_Lite_RP2040.h>
 
 #define HOST_NAME   "RP2040-Master-Controller"