+++ examples/SAM_DUE_WiFiNINA/defines.h
@@ -150,6 +150,11 @@
 
 /////////////////////////////////////////////
 
+// Optional, to use Board Name in Menu
+#define USING_BOARD_NAME                    true
+
+/////////////////////////////////////////////
+
 #include <WiFiManager_NINA_Lite_DUE.h>
 
 #define HOST_NAME   "DUE-Master-Controller"