+++ examples/STM32_WiFiNINA_MQTT/defines.h
@@ -164,6 +164,11 @@
 
 /////////////////////////////////////////////
 
+// Optional, to use Board Name in Menu
+#define USING_BOARD_NAME                    true
+
+/////////////////////////////////////////////
+
 #include "WiFiNINA_Pinout_Generic.h"
 
 #include <WiFiManager_NINA_Lite_STM32.h>